Fix SkippyBench eval reproducibility
